Emit an event called `momentumend` when momentum stops
In an application I'm building, I needed to know when the deceleration
had completed. To do so, I added this event so I could listen for it
from outside and do my things as needed. Nothing fancy, just something
I reckon could be useful for other people, hence the pull request.